cmd/compile: hoist some loop invariants
Conservatively hoist some loop invariants outside the loop.

Updates #15808
