+++ src/cmd/compile/internal/ssa/shortcircuit.go
@@ -473,41 +473,3 @@ func shortcircuitPhiPlan(b *Block, ctl *Value, cidx int, ti int64) func(*Value, 
 	// TODO: handle more cases; shortcircuit optimizations turn out to be reasonably high impact
 	return nil
 }
-
-// replaceUses replaces all uses of old in b with new.
-func (b *Block) replaceUses(old, new *Value) {
-	for _, v := range b.Values {
-		for i, a := range v.Args {
-			if a == old {
-				v.SetArg(i, new)
-			}
-		}
-	}
-	for i, v := range b.ControlValues() {
-		if v == old {
-			b.ReplaceControl(i, new)
-		}
-	}
-}
-
-// moveTo moves v to dst, adjusting the appropriate Block.Values slices.
-// The caller is responsible for ensuring that this is safe.
-// i is the index of v in v.Block.Values.
-func (v *Value) moveTo(dst *Block, i int) {
-	if dst.Func.scheduled {
-		v.Fatalf("moveTo after scheduling")
-	}
-	src := v.Block
-	if src.Values[i] != v {
-		v.Fatalf("moveTo bad index %d", v, i)
-	}
-	if src == dst {
-		return
-	}
-	v.Block = dst
-	dst.Values = append(dst.Values, v)
-	last := len(src.Values) - 1
-	src.Values[i] = src.Values[last]
-	src.Values[last] = nil
-	src.Values = src.Values[:last]
-}
